Move XException generic classes into Clash.XException.Internal
So we don't have to export:
  ( GShowX(..), GDeepErrorX(..), GHasUndefined(..), GEnsureSpine(..)
  , GNFDataX(..), Zero, One, ShowType(..), RnfArgs(..), NFDataX1(..) )
While also not generating haddock warnings.

Unfortunaly we also have to move some of the none internal stuff,
like ShowX, NFDataX,..
