+++ clash-prelude/clash-prelude.cabal
@@ -281,6 +281,7 @@ Library
                       Clash.Verification.Internal
 
                       Clash.XException
+                      Clash.XException.Internal
                       Clash.XException.TH
 
                       Clash.Xilinx.ClockGen
