+++ clash-prelude/src/Clash/Explicit/Prelude/Safe.hs
@@ -133,9 +133,7 @@ import Clash.Sized.RTree
 import Clash.Sized.Signed
 import Clash.Sized.Unsigned
 import Clash.Sized.Vector hiding (fromList, unsafeFromList)
-import Clash.XException hiding
-  ( GShowX(..), GDeepErrorX(..), GHasUndefined(..), GEnsureSpine(..)
-  , GNFDataX(..), Zero, One, ShowType(..), RnfArgs(..), NFDataX1(..) )
+import Clash.XException
 
 {- $setup
 >>> :set -XDataKinds
