+++ clash-prelude/src/Clash/XException.hs-boot
@@ -1,0 +1,3 @@
+module Clash.XException where
+
+isX :: a -> Either String a
